1. Change classes for SOLID 2. Removed redundant file reading 3. Added file existence check 4. Backup function moved to Remover class
