+++ ConsoleApp/ConsoleApp/FileHelpers/PathValidator.cs
